api/pyproject.toml
[project]
name = "FocaAi-backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.135.2",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "scalar-fastapi>=1.8.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]



[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "strict"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "factory-boy>=3.3.3",
    "faker>=40.13.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
